refactor(enums,error): 패키지 분리, LottoGame 클래스 패키지 위치 이동
