+++ src/test/java/lotto/shop/MoneyTest.java
@@ -3,7 +3,7 @@ package lotto.shop;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
-import lotto.common.ErrorMessage;
+import lotto.error.ErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
