+++ src/test/java/lotto/shop/ShopTest.java
@@ -2,13 +2,12 @@ package lotto.shop;
 
 import static org.assertj.core.api.Assertions.*;
 
-import lotto.common.ErrorMessage;
+import lotto.error.ErrorMessage;
 import lotto.lotto.Lotto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import javax.print.attribute.HashAttributeSet;
 import java.util.*;
 
 public class ShopTest {
